:truck: Move ui related files to ui package
