+++ app/src/main/java/com/changs/routesearch/ui/compose/SearchScreen.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.changs.routesearch.MapViewModel
+import com.changs.routesearch.ui.MapViewModel
 
 
 @OptIn(ExperimentalMaterial3Api::class)
