+++ app/src/main/res/layout/home_container.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/home_fragment_container"
-    android:name="com.changs.routesearch.HomeFragment"
+    android:name="com.changs.routesearch.ui.HomeFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"/>
